updated mandatory with new technologies

+++ bonus/src/parsing/parse_map.c
@@ -73,7 +73,7 @@ static void	parse_map_char(char map_char, int i, int j, t_game *g)
 
 static void	init_map_data(t_game *g)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < g->map.width * g->map.height)
@@ -85,8 +85,8 @@ static void	init_map_data(t_game *g)
 
 void	parse_map(char *map_str, t_game *g)
 {
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
 
 	fill_map_dimensions(map_str, g);
 	g->map.data = x_malloc(g->map.width * g->map.height * sizeof(int));
